Added HashMap to return fields specified by user and added unit test
User can pass the list of fields they want while loading the parser
The Rules are generated with these extra set of fields
A hashMap is returned which contains list of the specified fields and
the size of hashmap is always consistent whether a useragent string matches any rule or not

Invalid fields passed while loading the parser are not entertained and they are not present in HashMap object.

Existing Fixed fields are returned as it is for backward compatibility via defined methods.